+++ src/Symfony/Component/HttpKernel/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 2.4.0
 -----
 
+ * added event listeners for the session
  * added the KernelEvents::FINISH_REQUEST event
 
 2.3.0